+++ PilhaFat/src/view/Principal.java
@@ -13,7 +13,7 @@ public class Principal {
 		valor = Integer.parseInt(JOptionPane.showInputDialog("insira um valor de 0 a 10: "));
 		} while (valor <= 0 || valor >= 10);
 		
-		int rst = fc.Pilhar(valor);
+		long rst = fc.Pilhar(valor);
 		
 		System.out.println("O fatorial de " + valor + " é: " + rst);
 	}
